+++ src/main/java/be/baur/sdt/parser/SDTParseException.java
@@ -1,14 +1,14 @@
 package be.baur.sdt.parser;
 
 import be.baur.sda.Node;
-import be.baur.sda.ProcessingException;
+import be.baur.sda.NodeException;
 
 /**
  * An {@code SDTParseException} is thrown by an {@code SDTParser} if the SDS
  * syntax or semantics are violated.
  */
 @SuppressWarnings("serial")
-public class SDTParseException extends ProcessingException {
+public class SDTParseException extends NodeException {
 
 	/**
 	 * Creates an SDT parse exception with an error node and message.
@@ -25,10 +25,9 @@ public class SDTParseException extends ProcessingException {
 	 * Creates an SDT parse exception caused by another exception.
 	 * 
 	 * @param node    the node where an exception occurred
-	 * @param message an error message
 	 * @param cause   the exception causing this exception to be thrown
 	 */
-	public SDTParseException(Node node, String message, Throwable cause) {
+	public SDTParseException(Node node, Throwable cause) {
 		super(node, cause.getMessage()); initCause(cause);
 	}