+++ src/main/java/be/baur/sdt/parser/SDTParser.java
@@ -10,7 +10,8 @@ import org.jaxen.XPath;
 import be.baur.sda.DataNode;
 import be.baur.sda.Node;
 import be.baur.sda.SDA;
-import be.baur.sda.serialization.Parser;
+import be.baur.sda.io.Parser;
+import be.baur.sdt.SDT;
 import be.baur.sdt.transform.ChooseStatement;
 import be.baur.sdt.transform.CopyStatement;
 import be.baur.sdt.transform.ForEachStatement;
@@ -55,7 +56,7 @@ public final class SDTParser implements Parser<Transform> {
 	private static final String ATTRIBUTE_NOT_ALLOWED = "attribute '%s' is not allowed here";
 	private static final String KEYWORD_UNKNOWN = "keyword '%s' is unknown";
 	private static final String NODE_NAME_INVALID = "node name '%s' is invalid";
-	private static final String PARAMETER_REDECLARED = "parameter '%s' cannot be redeclared";
+	private static final String PARAMETER_REASSIGNED = "parameter '%s' cannot be reassigned";
 	private static final String PARAM_OVERWRITES_VARIABLE = "parameter '%s' cannot overwrite variable";
 	private static final String STATEMENT_EXPECTED = "'%s' statement expected";
 	private static final String ATTRIBUTE_EXPECTED_IN = "'%s' attribute expected in '%s'";
@@ -87,7 +88,7 @@ public final class SDTParser implements Parser<Transform> {
 		try {
 			sdt = SDA.parse(input);
 		} catch (Exception e) {
-			throw new SDTParseException(null, e.getMessage(), e);
+			throw new SDTParseException(null, e);
 		}
 		return parse(sdt);
 	}
@@ -271,7 +272,7 @@ public final class SDTParser implements Parser<Transform> {
 		if (value != null) // set the optional value expression
 			stat.setValueExpression(xpathFromNode(value));
 
-		for (Node node : sdt.find(n -> !n.getName().equals(Keyword.VALUE.tag))) // skip value attribute
+		for (Node node : sdt.getAll(n -> !n.getName().equals(Keyword.VALUE.tag))) // skip value attribute
 			stat.add(parseStatement((DataNode) node)); // parse and add child statements
 
 		return stat;
@@ -349,13 +350,13 @@ public final class SDTParser implements Parser<Transform> {
 		if (varname.isEmpty())
 			throw exception(sdt, STATEMENT_REQUIRES_VARIABLE, sdt.getName());
 
-		if (! VariableStatement.isVarName(varname))
+		if (! SDT.isVariableName(varname))
 			throw exception(sdt, VARIABLE_NAME_INVALID, varname);
 		
 		final Node parent = sdt.getParent();
 		
 		// find all declarations of a param with the specified name
-		List<Node> params = parent.find(n -> n.getName().equals(Keyword.PARAM.tag) 
+		List<Node> params = parent.getAll(n -> n.getName().equals(Keyword.PARAM.tag) 
 				&& ((DataNode) n).getValue().equals(varname));
 		
 		final boolean isParam = sdt.getName().equals(Keyword.PARAM.tag);
@@ -363,9 +364,9 @@ public final class SDTParser implements Parser<Transform> {
 		if ( isParam ) {
 			
 			if (params.size() > 1) // got more than one param
-				throw exception(params.get(1), PARAMETER_REDECLARED, varname);
+				throw exception(params.get(1), PARAMETER_REASSIGNED, varname);
 			
-			List<Node> vars = parent.findDescendant(n -> n.getName().equals(Keyword.VARIABLE.tag) 
+			List<Node> vars = parent.find(n -> n.getName().equals(Keyword.VARIABLE.tag) 
 					&& ((DataNode) n).getValue().equals(varname)); // find variables with this name
 			
 			if (vars.size() > 0) // got a variable with the same name
@@ -419,7 +420,7 @@ public final class SDTParser implements Parser<Transform> {
 		try {
 			xpath = new SDAXPath(node.getValue());
 		} catch (Exception e) {
-			throw new SDTParseException(node, e.getMessage(), e);
+			throw new SDTParseException(node, e);
 		}
 		return xpath;
 	}
@@ -487,7 +488,7 @@ public final class SDTParser implements Parser<Transform> {
 	 */
 	private static DataNode getAttribute(final DataNode sdt, Keyword att, Boolean required) throws SDTParseException {
 
-		List<DataNode> alist = sdt.find(n -> n.isLeaf() && n.getName().equals(att.tag));
+		List<DataNode> alist = sdt.getAll(n -> n.isLeaf() && n.getName().equals(att.tag));
 		int size = alist.size();
 
 		if (size == 0) {