+++ src/main/java/be/baur/sdt/xpath/SDAXPath.java
@@ -7,7 +7,9 @@ import org.jaxen.XPathFunctionContext;
 import be.baur.sdt.SDT;
 import be.baur.sdt.xpath.function.CompareNumberFunction;
 import be.baur.sdt.xpath.function.CompareStringFunction;
+import be.baur.sdt.xpath.function.DocumentNodeFunction;
 import be.baur.sdt.xpath.function.LeftFunction;
+import be.baur.sdt.xpath.function.ParseSDAFunction;
 import be.baur.sdt.xpath.function.RenderSDAFunction;
 import be.baur.sdt.xpath.function.RightFunction;
 import be.baur.sdt.xpath.function.StringJoinFunction;
@@ -22,7 +24,7 @@ import be.baur.sdt.xpath.function.TokenizeFunction;
  * example:
  *
  * <pre>
- * XPath path = new SDAXPath("/addressbook");
+ * XPath path = new SDAXPath("<i>expression</i>");
  * List results = path.selectNodes(node);
  * </pre>
  *
@@ -37,7 +39,9 @@ public class SDAXPath extends BaseXPath {
 		XPathFunctionContext ctx = (XPathFunctionContext) XPathFunctionContext.getInstance();
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "compare-number", new CompareNumberFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "compare-string", new CompareStringFunction());
+		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "document-node", new DocumentNodeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "left", new LeftFunction());
+		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "parse-sda", new ParseSDAFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "right", new RightFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "render-sda", new RenderSDAFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "tokenize", new TokenizeFunction());
@@ -48,6 +52,7 @@ public class SDAXPath extends BaseXPath {
 	 * Create a new <code>SDAXPath</code> from an XPath expression.
 	 *
 	 * @param expression the XPath expression
+	 * 
 	 * @throws JaxenException if there is a syntax error in the expression
 	 */
 	public SDAXPath(String expression) throws JaxenException {